# internal/admin/handlers.go
package admin

import (
	"encoding/json"
	"log/slog"
	"net"
	"net/http"
	"net/url"
	"sort"
	"strconv"
	"strings"
	"time"

	"github.com/gaetandev/waf/internal/jsonstrict"
	"github.com/gaetandev/waf/internal/signing"
	"github.com/gaetandev/waf/internal/storage"
	"github.com/gaetandev/waf/internal/trust"
)

// maxRequestBodyBytes borne le corps des requêtes admin (IPEntry,
// ConfigUpdate, GDPRErasureRequest : quelques centaines d'octets). Un corps
// plus grand échoue au décodage et reçoit le 400 de l'opération.
const maxRequestBodyBytes = 64 << 10

type errorResponse struct {
	Error   string `json:"error"`
	Message string `json:"message"`
}

type listResponse[T any] struct {
	Items []T `json:"items"`
	Total int `json:"total"`
}

type VisitorInfo struct {
	IPHash          string `json:"ip_hash"`
	Domain          string `json:"domain"`
	Score           int    `json:"score"`
	State           string `json:"state"`
	FirstSeen       string `json:"first_seen"`
	LastSeen        string `json:"last_seen"`
	ReqCount        int64  `json:"req_count"`
	ViolationCount  int    `json:"violation_count"`
	ChallengePassed bool   `json:"challenge_passed"`
	CircuitOpen     bool   `json:"circuit_open"`
}

type SecurityEventSummary struct {
	Timestamp  string `json:"timestamp"`
	RequestID  string `json:"request_id,omitempty"`
	IP         string `json:"ip"`
	Domain     string `json:"domain"`
	Method     string `json:"method,omitempty"`
	Path       string `json:"path,omitempty"`
	Action     string `json:"action"`
	Reason     string `json:"reason"`
	TrustScore int    `json:"trust_score"`
}

type WAFStats struct {
	UptimeSeconds       int64 `json:"uptime_seconds"`
	TotalRequests       int64 `json:"total_requests"`
	RequestsPassed      int64 `json:"requests_passed"`
	RequestsChallenged  int64 `json:"requests_challenged"`
	RequestsBlocked     int64 `json:"requests_blocked"`
	RequestsRateLimited int64 `json:"requests_rate_limited"`
	RequestsTarpitted   int64 `json:"requests_tarpitted"`
	ActiveVisitors      int   `json:"active_visitors"`
	TrustedVisitors     int   `json:"trusted_visitors"`
	MonitoredVisitors   int   `json:"monitored_visitors"`
	ChallengedVisitors  int   `json:"challenged_visitors"`
	BlockedVisitors     int   `json:"blocked_visitors"`
}

// route associe une opération du contrat specs/api/admin.openapi.yaml
// (specPath) à son motif ServeMux. Table unique : un test vérifie que chaque
// route servie a son opération dans le contrat (invariant #2).
type route struct {
	method   string
	pattern  string // chemin ServeMux (préfixe "/" final pour les paramètres)
	specPath string
	public   bool
	handler  http.HandlerFunc
}

func (s *Server) routeTable() []route {
	return []route{
		{method: http.MethodGet, pattern: "/waf/health", specPath: "/waf/health", public: true, handler: s.health},
		{method: http.MethodGet, pattern: "/waf/stats", specPath: "/waf/stats", handler: s.stats},
		{method: http.MethodGet, pattern: "/waf/admin/config", specPath: "/waf/admin/config", handler: s.getConfig},
		{method: http.MethodPatch, pattern: "/waf/admin/config", specPath: "/waf/admin/config", handler: s.patchConfig},
		{method: http.MethodGet, pattern: "/waf/admin/whitelist", specPath: "/waf/admin/whitelist", handler: s.getWhitelist},
		{method: http.MethodPost, pattern: "/waf/admin/whitelist", specPath: "/waf/admin/whitelist", handler: s.addWhitelist},
		{method: http.MethodDelete, pattern: "/waf/admin/whitelist/", specPath: "/waf/admin/whitelist/{ip}", handler: s.deleteWhitelist},
		{method: http.MethodGet, pattern: "/waf/admin/blacklist", specPath: "/waf/admin/blacklist", handler: s.getBlacklist},
		{method: http.MethodPost, pattern: "/waf/admin/blacklist", specPath: "/waf/admin/blacklist", handler: s.addBlacklist},
		{method: http.MethodDelete, pattern: "/waf/admin/blacklist/", specPath: "/waf/admin/blacklist/{ip}", handler: s.deleteBlacklist},
		{method: http.MethodGet, pattern: "/waf/admin/visitors", specPath: "/waf/admin/visitors", handler: s.listVisitors},
		{method: http.MethodGet, pattern: "/waf/admin/visitors/", specPath: "/waf/admin/visitors/{ip_hash}", handler: s.getVisitor},
		{method: http.MethodDelete, pattern: "/waf/admin/visitors/", specPath: "/waf/admin/visitors/{ip_hash}", handler: s.deleteVisitor},
		{method: http.MethodGet, pattern: "/waf/admin/events", specPath: "/waf/admin/events", handler: s.listEvents},
		{method: http.MethodGet, pattern: "/waf/admin/audit", specPath: "/waf/admin/audit", handler: s.listAudit},
		{method: http.MethodPost, pattern: "/waf/admin/gdpr/erase", specPath: "/waf/admin/gdpr/erase", handler: s.gdprErase},
	}
}

func (s *Server) routes() http.Handler {
	mux := http.NewServeMux()
	for _, r := range s.routeTable() {
		var handler http.Handler = r.handler
		if !r.public {
			handler = s.auth(handler)
		}
		mux.Handle(r.method+" "+r.pattern, limitBody(handler))
	}
	return mux
}

// limitBody plafonne la lecture du corps : sans borne, jsonstrict.Decode
// décodait en mémoire un flux de taille arbitraire.
func limitBody(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		r.Body = http.MaxBytesReader(w, r.Body, maxRequestBodyBytes)
		next.ServeHTTP(w, r)
	})
}

// gdprErase efface toutes les données d'un visiteur par son IP (droit à
// l'effacement, FR-28). L'IP est hashée pour retrouver l'entrée du store.
func (s *Server) gdprErase(w http.ResponseWriter, r *http.Request) {
	var payload struct {
		IP string `json:"ip"`
	}
	if err := jsonstrict.Decode(r.Body, &payload); err != nil || payload.IP == "" {
		writeJSON(w, http.StatusBadRequest, errorResponse{Error: "invalid_request", Message: "ip is required"})
		return
	}
	ipHash := trust.HashIP(payload.IP)
	_, existed := s.store.GetVisitor(ipHash)
	s.eraseVisitor(ipHash)
	s.record("gdpr_erase", ipHash, "erased")
	writeJSON(w, http.StatusOK, map[string]any{"erased": existed, "ip_hash": ipHash})
}

// eraseVisitor supprime le visiteur et tout l'état tenu hors du store de
// visiteurs (droit à l'effacement, FR-28).
func (s *Server) eraseVisitor(ipHash string) {
	s.store.DeleteVisitor(ipHash)
	for _, erase := range s.erasers {
		erase(ipHash)
	}
}

// record journalise une action d'administration (no-op si l'audit est désactivé).
func (s *Server) record(action string, target string, result string) {
	if s.trail != nil {
		s.trail.Record(action, target, result)
	}
}

func (s *Server) listAudit(w http.ResponseWriter, r *http.Request) {
	if s.trail == nil {
		writeJSON(w, http.StatusOK, listResponse[any]{Items: []any{}, Total: 0})
		return
	}
	writeJSON(w, http.StatusOK, paged(s.trail.List(), r))
}

func (s *Server) auth(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		ip := clientIP(r)
		// Anti-brute-force (FR-30) : verrouille l'IP après trop d'échecs.
		if s.brute != nil && s.brute.Limited(ip) {
			w.Header().Set("Retry-After", "300")
			writeJSON(w, http.StatusTooManyRequests, errorResponse{Error: "locked", Message: "Too many failed attempts"})
			return
		}
		if !bearerMatches(r.Header.Get("Authorization"), s.cfg.Admin.Token) {
			if s.brute != nil {
				s.brute.Record(ip)
			}
			writeJSON(w, http.StatusUnauthorized, errorResponse{Error: "unauthorized", Message: "Missing or invalid Bearer token"})
			return
		}
		next.ServeHTTP(w, r)
	})
}

// bearerMatches compare l'en-tête Authorization au token attendu, en temps
// constant longueur comprise.
func bearerMatches(header string, token string) bool {
	return signing.EqualSecret(header, "Bearer "+token)
}

func clientIP(r *http.Request) string {
	host, _, err := net.SplitHostPort(r.RemoteAddr)
	if err != nil {
		return r.RemoteAddr
	}
	return host
}

// degradationReporter est satisfait par le backend Redis : il sert son état
// local quand Redis est injoignable (ADR-021).
type degradationReporter interface {
	Degraded() bool
}

// health répond "degraded" quand le stockage partagé est en mode dégradé : le
// nœud sert, mais sur son seul état local. Le statut était toujours "ok".
func (s *Server) health(w http.ResponseWriter, _ *http.Request) {
	status := "ok"
	if reporter, ok := s.store.(degradationReporter); ok && reporter.Degraded() {
		status = "degraded"
	}
	writeJSON(w, http.StatusOK, map[string]any{
		"status":         status,
		"version":        s.cfg.Version,
		"uptime_seconds": int64(time.Since(s.startedAt).Seconds()),
	})
}

func (s *Server) stats(w http.ResponseWriter, _ *http.Request) {
	visitors := s.store.ListVisitors()
	stats := WAFStats{
		UptimeSeconds:  int64(time.Since(s.startedAt).Seconds()),
		ActiveVisitors: len(visitors),
	}
	s.counters.fill(&stats)
	for _, visitor := range visitors {
		switch s.scores.State(visitor.Score) {
		case "TRUSTED":
			stats.TrustedVisitors++
		case "MONITORED":
			stats.MonitoredVisitors++
		case "CHALLENGED":
			stats.ChallengedVisitors++
		case "BLOCKED":
			stats.BlockedVisitors++
		}
	}
	writeJSON(w, http.StatusOK, stats)
}

func (s *Server) getConfig(w http.ResponseWriter, _ *http.Request) {
	writeJSON(w, http.StatusOK, s.state.Config())
}

// patchConfig applique à chaud un sous-ensemble de la configuration.
//
// Le handler se contentait de vérifier le nom des sections et de journaliser
// « applied » : aucune valeur n'atteignait le rate limiter, le score manager ni
// le challenge.
func (s *Server) patchConfig(w http.ResponseWriter, r *http.Request) {
	if s.applyConfig == nil {
		writeJSON(w, http.StatusServiceUnavailable, errorResponse{Error: "hot_reload_unavailable", Message: "Runtime configuration cannot be updated on this instance"})
		return
	}
	var update ConfigUpdate
	if err := jsonstrict.Decode(r.Body, &update); err != nil {
		writeJSON(w, http.StatusBadRequest, errorResponse{Error: "invalid_config_update", Message: "Invalid JSON body or unsupported config field"})
		return
	}
	next, updated, err := s.state.ApplyConfigUpdate(update)
	if err != nil {
		writeJSON(w, http.StatusBadRequest, errorResponse{Error: "invalid_config_update", Message: err.Error()})
		return
	}
	if len(updated) == 0 {
		writeJSON(w, http.StatusBadRequest, errorResponse{Error: "invalid_config_update", Message: "No supported field to update"})
		return
	}
	s.applyConfig(next)
	s.record("config_patch", strings.Join(updated, ","), "applied")
	writeJSON(w, http.StatusOK, map[string]any{"updated_fields": updated})
}

func (s *Server) getWhitelist(w http.ResponseWriter, r *http.Request) {
	writeJSON(w, http.StatusOK, paged(s.state.ListWhitelist(), r))
}

func (s *Server) addWhitelist(w http.ResponseWriter, r *http.Request) {
	s.addIPEntry(w, r, true)
}

func (s *Server) deleteWhitelist(w http.ResponseWriter, r *http.Request) {
	target := pathTail(r, "/waf/admin/whitelist/")
	found, err := s.state.RemoveWhitelist(target)
	if !found {
		http.NotFound(w, r)
		return
	}
	if err != nil {
		writeSyncFailure(w, err)
		return
	}
	s.record("remove_whitelist", target, "removed")
	w.WriteHeader(http.StatusNoContent)
}

func (s *Server) getBlacklist(w http.ResponseWriter, r *http.Request) {
	writeJSON(w, http.StatusOK, paged(s.state.ListBlacklist(), r))
}

func (s *Server) addBlacklist(w http.ResponseWriter, r *http.Request) {
	s.addIPEntry(w, r, false)
}

func (s *Server) deleteBlacklist(w http.ResponseWriter, r *http.Request) {
	target := pathTail(r, "/waf/admin/blacklist/")
	found, err := s.state.RemoveBlacklist(target)
	if !found {
		http.NotFound(w, r)
		return
	}
	if err != nil {
		writeSyncFailure(w, err)
		return
	}
	s.record("remove_blacklist", target, "removed")
	w.WriteHeader(http.StatusNoContent)
}

// writeSyncFailure signale que les règles appliquées n'ont pas pu suivre une
// suppression : l'entrée est restaurée, rien n'a changé.
func writeSyncFailure(w http.ResponseWriter, err error) {
	slog.Error("access rules sync failed: entry restored", "error", err)
	writeJSON(w, http.StatusInternalServerError, errorResponse{Error: "access_rules_sync_failed", Message: "Access rules could not be updated; the entry was kept"})
}

func (s *Server) addIPEntry(w http.ResponseWriter, r *http.Request, whitelist bool) {
	var entry IPEntry
	if err := jsonstrict.Decode(r.Body, &entry); err != nil {
		writeJSON(w, http.StatusBadRequest, errorResponse{Error: "invalid_ip_entry", Message: "Invalid JSON body"})
		return
	}
	var created IPEntry
	var exists bool
	var err error
	if whitelist {
		created, exists, err = s.state.AddWhitelist(entry)
	} else {
		created, exists, err = s.state.AddBlacklist(entry)
	}
	if err != nil {
		writeJSON(w, http.StatusBadRequest, errorResponse{Error: "invalid_ip", Message: "IP or CIDR is invalid"})
		return
	}
	if exists {
		writeJSON(w, http.StatusConflict, errorResponse{Error: "already_exists", Message: "Entry already exists"})
		return
	}
	action := "add_blacklist"
	if whitelist {
		action = "add_whitelist"
	} else if s.onBlacklist != nil {
		s.onBlacklist(created.IP)
	}
	s.record(action, entry.IP, "created")
	writeJSON(w, http.StatusCreated, created)
}

func (s *Server) listVisitors(w http.ResponseWriter, r *http.Request) {
	visitors := visitorInfos(s.store.ListVisitors(), s.scores)
	domain := r.URL.Query().Get("domain")
	state := r.URL.Query().Get("state")
	filtered := visitors[:0]
	for _, visitor := range visitors {
		if domain != "" && visitor.Domain != domain {
			continue
		}
		if state != "" && visitor.State != state {
			continue
		}
		filtered = append(filtered, visitor)
	}
	sortVisitors(filtered, r.URL.Query().Get("sort"))
	writeJSON(w, http.StatusOK, paged(filtered, r))
}

func (s *Server) getVisitor(w http.ResponseWriter, r *http.Request) {
	ipHash := pathTail(r, "/waf/admin/visitors/")
	visitor, ok := s.store.GetVisitor(ipHash)
	if !ok {
		http.NotFound(w, r)
		return
	}
	writeJSON(w, http.StatusOK, visitorInfo(*visitor, s.scores))
}

func (s *Server) deleteVisitor(w http.ResponseWriter, r *http.Request) {
	ipHash := pathTail(r, "/waf/admin/visitors/")
	if _, ok := s.store.GetVisitor(ipHash); !ok {
		http.NotFound(w, r)
		return
	}
	s.eraseVisitor(ipHash)
	s.record("reset_visitor", ipHash, "reset")
	w.WriteHeader(http.StatusNoContent)
}

// listEvents sert les événements de sécurité récents. Le flux était toujours
// vide : rien n'écrivait dans le tampon d'événements.
func (s *Server) listEvents(w http.ResponseWriter, r *http.Request) {
	var since time.Time
	if raw := r.URL.Query().Get("since"); raw != "" {
		parsed, err := time.Parse(time.RFC3339, raw)
		if err != nil {
			writeJSON(w, http.StatusBadRequest, errorResponse{Error: "invalid_since", Message: "since must be an RFC 3339 date-time"})
			return
		}
		since = parsed
	}
	events := s.events.Recent(since)
	domain := r.URL.Query().Get("domain")
	action := r.URL.Query().Get("action")
	filtered := events[:0]
	for _, event := range events {
		if domain != "" && event.Domain != domain {
			continue
		}
		if action != "" && event.Action != action {
			continue
		}
		filtered = append(filtered, event)
	}
	writeJSON(w, http.StatusOK, paged(filtered, r))
}

func visitorInfos(visitors []storage.VisitorState, scores *trust.ScoreManager) []VisitorInfo {
	items := make([]VisitorInfo, 0, len(visitors))
	for _, visitor := range visitors {
		items = append(items, visitorInfo(visitor, scores))
	}
	return items
}

func visitorInfo(visitor storage.VisitorState, scores *trust.ScoreManager) VisitorInfo {
	return VisitorInfo{
		IPHash:          visitor.IPHash,
		Domain:          visitor.Domain,
		Score:           visitor.Score,
		State:           scores.State(visitor.Score),
		FirstSeen:       visitor.FirstSeen.UTC().Format(time.RFC3339),
		LastSeen:        visitor.LastSeen.UTC().Format(time.RFC3339),
		ReqCount:        visitor.ReqCount,
		ViolationCount:  visitor.ViolationCount,
		ChallengePassed: visitor.ChallengePassed,
		CircuitOpen:     visitor.CircuitOpen,
	}
}

func sortVisitors(visitors []VisitorInfo, sortBy string) {
	sort.Slice(visitors, func(i, j int) bool {
		switch sortBy {
		case "score_asc":
			return visitors[i].Score < visitors[j].Score
		case "score_desc":
			return visitors[i].Score > visitors[j].Score
		default:
			return visitors[i].LastSeen > visitors[j].LastSeen
		}
	})
}

func paged[T any](items []T, r *http.Request) listResponse[T] {
	total := len(items)
	page := queryInt(r, "page", 1)
	limit := min(queryInt(r, "limit", 50), 1000)
	start := (page - 1) * limit
	if start >= total {
		return listResponse[T]{Items: []T{}, Total: total}
	}
	end := min(start+limit, total)
	return listResponse[T]{Items: items[start:end], Total: total}
}

func queryInt(r *http.Request, name string, defaultValue int) int {
	value, err := strconv.Atoi(r.URL.Query().Get(name))
	if err != nil || value < 1 {
		return defaultValue
	}
	return value
}

func pathTail(r *http.Request, prefix string) string {
	escaped := strings.TrimPrefix(r.URL.EscapedPath(), prefix)
	value, err := url.PathUnescape(escaped)
	if err != nil {
		return escaped
	}
	return value
}

func writeJSON(w http.ResponseWriter, status int, payload any) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	_ = json.NewEncoder(w).Encode(payload)
}