# specs/api/admin.openapi.yaml
openapi: "3.1.0"
info:
  title: WAF Admin API
  description: |
    API d'administration du WAF Anti-DDoS/Anti-Bot.
    Exposée sur un port séparé (défaut: 9090), protégée par Bearer token.
    Ne jamais exposer ce port publiquement.
    Tout corps de requête est borné à 64 Kio : au-delà, le décodage échoue et
    l'opération répond son 400.
  version: 1.4.1
  contact:
    name: GaetanDev
    url: https://firewall.gaetandev.fr
servers:
  - url: http://127.0.0.1:9090
    description: Admin API (port privé)
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [error, message]
      properties:
        error:
          type: string
          description: Code d'erreur machine-readable
        message:
          type: string
          description: Description humaine de l'erreur
...
    WAFStats:
      type: object
      additionalProperties: false
      description: |
        Compteurs requests_* : requêtes traitées par ce nœud depuis son
        démarrage, par action journalisée. Les compteurs *_visitors portent sur
        les visiteurs suivis par le store.
      properties:
        uptime_seconds:
          type: integer
        total_requests:
          type: integer
          description: Toutes les requêtes journalisées depuis le démarrage
        requests_passed:
          type: integer
          description: Action PASS
        requests_challenged:
          type: integer
          description: Action CHALLENGE (page de challenge servie)
        requests_blocked:
          type: integer
          description: Actions BLOCK, CIRCUIT_BREAK et HONEYPOT
        requests_rate_limited:
          type: integer
          description: Action RATE_LIMIT
        requests_tarpitted:
          type: integer
          description: Action TARPIT (réponse servie par le tarpit, FR-15)
        active_visitors:
          type: integer
        trusted_visitors:
          type: integer
        monitored_visitors:
          type: integer
        challenged_visitors:
          type: integer
        blocked_visitors:
          type: integer
...
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 50
...
  responses:
    Locked:
      description: >-
        IP verrouillée après self_protection.admin_max_failures échecs
        d'authentification (FR-30), pendant self_protection.admin_lockout,
        token valide compris.
      headers:
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "locked"
            message: "Too many failed attempts"
    Conflict:
      description: Entrée déjà existante
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "already_exists"
            message: "Entry already exists"
...
    Unauthorized:
      description: Token Bearer manquant ou invalide
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "unauthorized"
            message: "Missing or invalid Bearer token"
paths:
...
  /waf/stats:
    get:
      operationId: getStats
      summary: Statistiques globales du WAF
      description: Retourne les compteurs agrégés du WAF (requêtes, blocages, challenges, visiteurs actifs par état).
      tags: [System]
      responses:
        "200":
          description: Statistiques courantes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WAFStats"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/Locked"
  /waf/admin/config:
    get:
      operationId: getConfig
      summary: Récupérer la configuration active
      description: >-
        Retourne la configuration courante du WAF, secrets masqués par "***" :
        challenge.secret_key, admin.token, origin_protection.secret,
        threat_intel.abuseipdb.api_key, storage.redis.password et
        alerting.webhooks[].url (une URL de webhook porte son jeton d'accès).
      tags: [Configuration]
      responses:
        "200":
          description: Configuration courante (secrets masqués)
          content:
            application/json:
              schema:
                description: Voir config.schema.json — secrets remplacés par "***"
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/Locked"

    patch:
      operationId: updateConfig
      summary: Mettre à jour des paramètres de configuration (hot-reload)
      description: Applique à chaud un sous-ensemble de paramètres de configuration sans redémarrer le WAF. Retourne la liste des champs modifiés, en notation pointée (ex. "rate_limit.burst").
      tags: [Configuration]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigUpdate"
      responses:
        "200":
          description: Configuration mise à jour
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [updated_fields]
                properties:
                  updated_fields:
                    type: array
                    items:
                      type: string
                    examples:
                      - ["rate_limit.burst", "trust.challenge_threshold"]
        "400":
          description: Champ inconnu, aucune modification, ou configuration résultante invalide (rien n'est appliqué)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/Locked"
        "503":
          description: Hot-reload indisponible sur cette instance (aucune modification appliquée)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /waf/admin/whitelist:
    get:
      operationId: getWhitelist
      summary: Lister les entrées de la whitelist
      description: Retourne la liste paginée des IPs et CIDRs autorisés (bypass total des protections).
      tags: [Access Control]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Liste des IPs/CIDRs en whitelist
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items, total]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/IPEntry"
                  total:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/Locked"

    post:
      operationId: addToWhitelist
      summary: Ajouter une IP ou CIDR en whitelist
      description: Ajoute une IP ou un CIDR à la whitelist. L'entrée prend effet immédiatement (hot-reload).
      tags: [Access Control]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IPEntry"
      responses:
        "201":
          description: Entrée ajoutée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IPEntry"
        "400":
          description: IP/CIDR invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/Locked"
  /waf/admin/whitelist/{ip}:
    delete:
      operationId: removeFromWhitelist
      summary: Supprimer une IP de la whitelist
      description: Retire une IP ou un CIDR de la whitelist. La suppression prend effet immédiatement.
      tags: [Access Control]
      parameters:
        - $ref: "#/components/parameters/IP"
      responses:
        "204":
          description: Entrée supprimée
        "404":
          description: IP non trouvée
        "500":
          $ref: "#/components/responses/AccessRulesSyncFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/Locked"
  /waf/admin/blacklist:
    get:
      operationId: getBlacklist
      summary: Lister les entrées de la blacklist
      description: Retourne la liste paginée des IPs et CIDRs bloqués (réponse HTTP 403).
      tags: [Access Control]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Liste des IPs/CIDRs en blacklist
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items, total]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/IPEntry"
                  total:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/Locked"

    post:
      operationId: addToBlacklist
      summary: Ajouter une IP ou CIDR en blacklist
      description: Ajoute une IP ou un CIDR à la blacklist. Les requêtes correspondantes sont rejetées en HTTP 403 immédiatement.
      tags: [Access Control]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IPEntry"
      responses:
        "201":
          description: Entrée ajoutée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IPEntry"
        "400":
          description: IP/CIDR invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/Locked"
  /waf/admin/blacklist/{ip}:
    delete:
      operationId: removeFromBlacklist
      summary: Supprimer une IP de la blacklist
      description: Retire une IP ou un CIDR de la blacklist. La suppression prend effet immédiatement.
      tags: [Access Control]
      parameters:
        - $ref: "#/components/parameters/IP"
      responses:
        "204":
          description: Entrée supprimée
        "404":
          description: IP non trouvée
        "500":
          $ref: "#/components/responses/AccessRulesSyncFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/Locked"
  /waf/admin/visitors:
    get:
      operationId: listVisitors
      summary: Lister les visiteurs actifs avec leur score
      description: Retourne la liste paginée des visiteurs suivis, filtrable par domaine et état, triable par score ou dernière activité.
      tags: [Visitors]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Domain"
        - name: state
          in: query
          schema:
            type: string
            enum: [TRUSTED, MONITORED, CHALLENGED, BLOCKED]
        - name: sort
          in: query
          schema:
            type: string
            enum: [score_asc, score_desc, last_seen_desc]
            default: last_seen_desc
      responses:
        "200":
          description: Liste des visiteurs actifs
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items, total]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/VisitorInfo"
                  total:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/Locked"
  /waf/admin/visitors/{ip_hash}:
    get:
      operationId: getVisitor
      summary: Détails d'un visiteur par hash IP
      description: Retourne le score, l'état et les métadonnées d'un visiteur identifié par son hash IP.
      tags: [Visitors]
      parameters:
        - name: ip_hash
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Détails du visiteur
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitorInfo"
        "404":
          description: Visiteur non trouvé
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/Locked"

    delete:
      operationId: resetVisitor
      summary: Réinitialiser le score et l'état d'un visiteur
      description: Réinitialise le score au score initial et efface l'état du visiteur identifié par son hash IP.
      tags: [Visitors]
      parameters:
        - name: ip_hash
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: État du visiteur réinitialisé
        "404":
          description: Visiteur non trouvé
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/Locked"
  /waf/admin/events:
    get:
      operationId: listEvents
      summary: Lister les événements de sécurité récents
      description: |
        Retourne les événements de sécurité récents, du plus récent au plus
        ancien, filtrables par domaine, action et date. Le flux est alimenté par
        le middleware de journalisation : seules les décisions de mitigation
        sont retenues (PASS exclu), dans la limite de 10 000 événements et de
        24 h. L'IP est anonymisée si gdpr.anonymize_ip est actif.
      tags: [Events]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Domain"
        - name: action
          in: query
          schema:
            type: string
            enum: [PASS, CHALLENGE, BLOCK, RATE_LIMIT, CIRCUIT_BREAK, HONEYPOT, TARPIT]
        - name: since
          in: query
          schema:
            type: string
            format: date-time
          description: Filtrer les événements depuis cette date
      responses:
        "200":
          description: Liste des événements récents (max 24h en mémoire)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items, total]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/SecurityEventSummary"
                  total:
                    type: integer
        "400":
          description: Paramètre since invalide (date-time RFC 3339 attendue)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/Locked"
  /waf/admin/audit:
    get:
      operationId: listAuditEntries
      summary: Consulter le journal d'audit des actions d'administration
      description: |
        Retourne le journal d'audit (FR-27), de la plus ancienne à la plus
        récente entrée, borné à audit.max_entries (rotation FIFO). Liste vide
        si audit.enabled est faux. Journal append-only : aucune méthode ne le
        modifie (DELETE répond 405).
      tags: [Audit]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Entrées d'audit
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items, total]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuditEntry"
                  total:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/Locked"
  /waf/admin/gdpr/erase:
    post:
      operationId: eraseVisitorData
      summary: Effacer les données d'un visiteur (droit à l'effacement)
      description: |
        Supprime l'état du visiteur identifié par son IP (RGPD article 17,
        FR-28) et journalise l'action (gdpr_erase) dans l'audit.
      tags: [GDPR]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GDPRErasureRequest"
      responses:
        "200":
          description: Données effacées (ou absentes)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GDPRErasureResult"
        "400":
          description: Corps invalide ou IP manquante
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "invalid_request"
                message: "ip is required"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/Locked"
...